Adds js reorganizes test
