+++ problems/_001/compiled/euler_1.hi
