+++ problems/_001/euler_001.hs
