+++ problems/_002/compiled/euler_002.hi
