+++ problems/_002/euler_002.hs
